requirements.txt
dj-database-url==0.5.0
Django==2.0.13
django-google-charts==0.1.1
django-tinymce==2.8.0
django-uuslug==1.1.8
gunicorn==19.9.0
Pillow==6.0.0
psycopg2==2.8.2
python-slugify==3.0.2
pytz==2019.1
slugify==0.0.1
sqlparse==0.3.0
text-unidecode==1.2
virtualenv==16.6.0
virtualenvwrapper-win==1.2.5
whitenoise==4.1.2
django-qsstats-magic==1.0.0
django-resized==0.3.9